Zoomed out combat camera, reworking combat movement
I decided to zoom out the camera for combat because some attack animations look way better when you can see the character's whole body. Zooming out the camera exposes the legs, which shows how crappy the animations for strafing look. So I'm reworking strafing and finding new animations for it.