Added great sword block
I added the great sword block animation and repositioned the sword so it looks right.

I also stretched out the sword. This is gonna mess up the whole sheath stuff we have going on - so we'll have to adjust the sheath size and positioning too.  We can do that later though.

Right now we should figure out dodging animations for great sword.