Added dodge animation, camera transition smoothing, gotta work on attacks next
I added the great sword jump as the dodge animation. It's not great but it's serviceable. We can dress it up with foot dust and other stuff. At least it's fast and responsive.

I also worked on smoothing out the transition in and out of the combat camera, specificially smoothing out the change in horizontalPosOffset.

Now we need to get started on attacks. I think it should be kind of straightforward. We need to add ROOT MOTION animations for attacks and use MatchTarget position.

We find out the z displacement of the attack animation. We find out the distance from you to the closest position to the enemy. We decide on a certain range of distance that we can round. Like after the attack if we are just 1 unit away from the enemy then we can round up 1 unit.

We see if the animation displacement  is greater than or in a roundable difference up to the enemy player distance. If it is, then through the animation you move to the desired position in front of the enemy. If it is not,  then you just move the distance of the root motion attack normally, you don't use match target.