Added target matching, needs work still
I've added basic target matching functionality but it still needs a ton of work and tweaking.

What's also really annoying me now is the enemy's reactions. Both the timing and the animations. We should pair up the animations with each attack like we did before, and we should time it correctly. Also we should move the enemy back after an attack so the next attack can move forward. I think we should fix this stuff first and then go back to target matching.