Fixed rolling camera angle and camera angle unwrapping
When rolling, the camera looks in the line from the player's position to the enemy's position. I did this by just putting a new transform as Dev's child, making it look at the enemy, and using that transform's eulerangles for the camera.

I fixed the camera unwrapping by making sure that the next angle is not wrapped around from the previous angle. For example, if the angle used to be 1f, it can't be 359f now! It would have to be -1f now. I thought this might cause problems if the number gets too big, but that will only occur if someone spins around thousands of times in the same direction in the same combat scenario.