Fixed game freezing, tweaked rolling distance
Turned out the issue was that the inCombatMode function had "ShieldDraw" in it, which isn't even an animator variable. Taking that part out removed the freezing.

I kept the 0.8 animation speed for rolling, but I change the movement multiplier for rolling from 2 to 1.2 so that we don't move that much and you're encouraged to roll more.