Got match target method to fully run
The match target update method would exit early because we were checking if the tag of the animator is attacking. The tag of the animator is the tag of the dev gameobject, which is unset.

What we wanted to be doing is checking the tag of the current state in the animator. Once I made this change, the method ran fully.

I tweaked some of the timing and the desired distances. It's not perfect yet but now that the metho is actually running, we're on the right track.

As a side note, I feel that it is necessary for the enemy to move back. Even if we get match target to work really well, it's not going to look good if the enemy doesn't move back significantly.

Right now the moving back helps, except what we need is an animation showing him moving back -- not just a translation backwards.