Significantly improved matchtarget
Match target works well now for attacks. The stickiness is decent and the match end times are good too.

The problem now is that there is a huge visual difference between if we are doing match target or not.

If the character is too far to do match target, and we don't do it -- then the animation barely moves.

If we are doing match target, then the same animation moves a lot.

We have to make both cases move alot, even if they don't attack.