Added extra velocity even when we miss
Now every attack does match target, even if it hits or not. This way, every attack has velocity.

The exception is when you are super close. When you are super close, you hardly move. In the future we could find some improvement for this, but for now match target is pretty great.

We need to work on enemy backward animations now.